cmd/containerboot,cmd/k8s-operator: enable IPv6 for fqdn egress proxies
Don't skip installing egress forwarding rules for IPv6, and set headless
services ipFamilyPolicy to PreferDualStack to optionally enable both IP
families when possible. Note that even with PreferDualStack set, testing
a dual-stack GKE cluster with the default DNS setup of kube-dns did not
correctly set both A and AAAA records for the headless service, and
instead only did so when switching the cluster DNS to Cloud DNS. For
both IPv4 and IPv6 to work simultaneously in a dual-stack cluster, we
require headless services to return both A and AAAA records.

Fixes #12215

Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>