ipn/ipnlocal: reload some prefs changes on tailscaled configfile changes for running system
This is the first part of work towards making it possible to reload
tailscaled config in containers, so that the runner does not
need to restart the container on configfile changes.

This change ensures that if values of '.AcceptRoutes' and
'.AdvertizeRoutes' have changed when reload local API endpoint is
called, the change will be applied for a LocalBackend in 'Running' state.
These specific fields where chosen because those are the values that
can change by user action in kube operator proxies.
We can then gradually allow more values to be reloaded as we make
the configfile easier to use in any containerized environment.

Updates tailscale/tailscale#13032

Signed-off-by: Irbe Krumina <irbe@tailscale.com>