go.mod: bump wireguard-go to 6.6.44 mitigation test
Linux 6.6.44 introduced a regression in GSO handling that spuriously
returns EINVAL for tail packets shorter than gso_size. The referenced
wireguard-go patch works around this by not including packets matching
this shape in the offload, which should be rare, but are dependent on
traffic patterns.

This patch tests out fine in my lab, creating a branch for original
reporters to also test with.

Updates #13041

Signed-off-by: James Tucker <james@tailscale.com>