wgengine/magicsock: fix the freshness of endpoints in callmemaybe
Endpoint storage and state updates are all over the place so the full
story is hard to track, but explicitly callmemaybe is executed by a
callback that is triggered before the write of new endpoints into the
storage that CallMeMaybe sends read from. This is now reordered, so that
CallMeMaybe should always observe the most recent endpoints.

Signed-off-by: James Tucker <james@tailscale.com>